* Tagged 0.3 alpha 4 release, final
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@450 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ luascript.cpp
@@ -2177,7 +2177,7 @@ int32_t LuaScriptInterface::internalGetPlayerInfo(lua_State* L, PlayerInfo_t inf
 				value = player->getPlayerInfo(PLAYERINFO_SOUL);
 				break;
 			case PlayerInfoFreeCap:
-				value = player->getFreeCapacity();
+				value = (int64_t)player->getFreeCapacity();
 				break;
 			case PlayerInfoGuildId:
 				value = player->getGuildId();
